feat(indexer): add supports for wiki-links and frontmatter
